change decompilation modal to a tab
